src/requirements-frozen.txt
asgiref==3.11.1
attrs==25.4.0
django==6.0.2
django-cors-headers==4.9.0
django-stubs==5.2.9
django-stubs-ext==5.2.9
djangorestframework==3.16.1
djangorestframework-stubs==3.16.8
drf-spectacular==0.29.0
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
pyyaml==6.0.3
referencing==0.37.0
rpds-py==0.30.0
sqlparse==0.5.5
types-pyyaml==6.0.12.20250915
typing-extensions==4.15.0
uritemplate==4.2.0
